Reverted CSVImport bulk updates to individually save recipients and recipientattributes; improved error handling; fixed fetch of recipientgroup in revert()

+++ manager/utils.py
@@ -45,6 +45,7 @@ class CSVImport:
     subprocess = None
     update_factor = 1
     remove_stale = False
+    new_group = False
 
     def __init__(self, csv_file, recipient_group_name, skip_first_row, column_order, subprocess, remove_stale=False, stderr=None):
         if csv_file:
@@ -76,14 +77,13 @@ class CSVImport:
             print('email is a required column for import')
             raise Exception('email is a required column for import')
 
-        new_group = False
         group = None
         try:
             group = RecipientGroup.objects.get(name=self.recipient_group_name)
         except RecipientGroup.DoesNotExist:
             log.info('Recipient group does not exist. Creating...')
+            self.new_group = True
             group = RecipientGroup(name=self.recipient_group_name)
-            new_group = True
             group.save()
 
         if self.remove_stale:
@@ -92,8 +92,7 @@ class CSVImport:
         if self.subprocess:
             self.tracker = SubprocessStatus.objects.get(pk=self.subprocess)
             self.tracker.total_units = self.get_line_count()
-            if new_group:
-                self.tracker.success_url = reverse('manager-recipientgroup-update', kwargs={'pk': group.pk})
+            self.tracker.success_url = reverse('manager-recipientgroup-update', kwargs={'pk': group.pk})
             self.tracker.save()
 
             # Update the update_factor to prevent a save every time it loops through
@@ -129,8 +128,6 @@ class CSVImport:
         attribute_creates = []
         attribute_updates = []
 
-        created_objs = []
-
         for idx, row in enumerate(csv_reader):
             if self.skip_first_row and idx == 0: continue
 
@@ -141,13 +138,14 @@ class CSVImport:
                 first_name = row['first_name'].strip() if 'first_name' in csv_reader.fieldnames else None
                 last_name = row['last_name'].strip() if 'last_name' in csv_reader.fieldnames else None
                 preferred_name = row['preferred_name'].strip() if 'preferred_name' in csv_reader.fieldnames else None
-            except IndexError:
+            except (IndexError, AttributeError):
                 self.revert()
-                self.update_status("Error", f'There is a malformed row at line {row_num}')
+                self.update_status("Error", f'There is a malformed row at line {row_num} of the provided CSV.', row_num)
                 raise Exception(f'There is a malformed row at line {row_num}')
             else:
                 if email_address == '':
-                    log.error(('Empty email address at line %d' % row_num))
+                    log.error(f'Skipping row with empty email address at line {row_num}')
+                    continue
                 else:
                     try:
                         recipient = Recipient.objects.get(email_address=email_address)
@@ -155,13 +153,23 @@ class CSVImport:
                         processed_emails.append(email_address)
                     except:
                         recipient = Recipient(
-                                email_address=email_address
+                            email_address=email_address
                         )
 
                         # Protect against dupe insert
                         if email_address not in processed_emails:
                             processed_emails.append(email_address)
                             recipient_creates.append(recipient)
+                        else:
+                            log.error(f'Skipping row with duplicate email address {email_address} at line {row_num}')
+                            continue
+
+                    try:
+                        recipient.save()
+                    except Exception as e:
+                        self.revert()
+                        self.update_status("Error", f'Error saving recipient at line {row_num} of the provided CSV', row_num)
+                        raise Exception(f'Error saving recipient at line {row_num}: {str(e)}')
 
 
                     if first_name is not None:
@@ -177,6 +185,11 @@ class CSVImport:
                             )
                             attribute_creates.append(attribute_first_name)
 
+                        try:
+                            attribute_first_name.save()
+                        except Exception as e:
+                            log.error(f'Error saving recipient attribute First Name at line {row_num}: {str(e)}')
+
 
                     if last_name is not None:
                         try:
@@ -190,6 +203,11 @@ class CSVImport:
                                 value = last_name
                             )
                             attribute_creates.append(attribute_last_name)
+
+                        try:
+                            attribute_last_name.save()
+                        except Exception as e:
+                            log.error(f'Error saving recipient attribute Last Name at line {row_num}: {str(e)}')
 
 
                     if preferred_name is not None:
@@ -205,22 +223,21 @@ class CSVImport:
                                 value = preferred_name
                             )
                             attribute_creates.append(attribute_preferred_name)
+
+                        try:
+                            attribute_preferred_name.save()
+                        except Exception as e:
+                            log.error(f'Error saving recipient attribute Preferred Name at line {row_num}: {str(e)}')
             # Increment
             self.update_status("In Progress", "", row_num)
 
-        created_objs = Recipient.objects.bulk_create(recipient_creates, batch_size=100)
-
-        RecipientAttribute.objects.bulk_create(attribute_creates, batch_size=100)
-        RecipientAttribute.objects.bulk_update(attribute_updates, ['value'], batch_size=100)
-
-        if group is not None:
-            try:
-                all_objs = created_objs + recipient_updates
-                group.recipients.add(*all_objs)
-            except Exception as e:
-                pass
-
-        self.update_status("Completed", "", self.tracker.total_units)
+        try:
+            all_objs = recipient_creates + recipient_updates
+            group.recipients.add(*all_objs)
+            self.update_status("Completed", "", self.tracker.total_units)
+        except Exception as e:
+            self.update_status("Error", f'Error adding recipients to recipient group', self.tracker.total_units)
+            raise Exception(f'Error adding recipients to recipient group: {str(e)}')
 
         if self.subprocess:
             self.delete_file(self.csv_file.name)
@@ -239,6 +256,7 @@ class CSVImport:
             self.tracker.status = status
             self.tracker.error = error
             self.tracker.current_unit = current_unit
+            self.tracker.success_url = reverse('manager-recipients-csv-import')
             self.tracker.save()
 
     def delete_file(self, filename):
@@ -248,12 +266,15 @@ class CSVImport:
         self.tracker.delete()
 
     def revert(self):
-        try:
-            recipient_group = RecipientAttribute.objects.get(name=self.recipient_group_name)
-        except:
-            return
-        else:
-            recipient_group.delete()
+        if self.new_group:
+            try:
+                # TODO should we even be doing this?
+                recipient_group = RecipientGroup.objects.get(name=self.recipient_group_name)
+            except:
+                return
+            else:
+                recipient_group.delete()
+        return
 
     def get_line_count(self):
         return sum(1 for line in self.csv_file)
